move date_builder.dart into widgets directory
